Add 'BulkUpdateQuerySet' that fixes #47.
